Re-enable sign-compare and double-promotion warnings (#3629)
* Fix unsigned comparison and implicit double conversion

* Eliminate sign/unsigned compare warnings, implicit double conversion

* Missed type

* Neww int type

* omp compromise

* style

+++ src/ngraph/frontend/onnx_import/op/onehot.cpp
@@ -56,7 +56,8 @@ namespace ngraph
                         axis += indices_shape.size() + 1;
                     }
 
-                    ASSERT_VALID_ARGUMENT(node, (axis >= 0) && (axis <= indices_shape.size()))
+                    ASSERT_VALID_ARGUMENT(
+                        node, (axis >= 0) && (axis <= static_cast<int64_t>(indices_shape.size())))
                         << "invalid 'axis' attribute: "
                         << node.get_attribute_value<std::int64_t>("axis", -1);
